Fix: Use ULong for encoding on all 64-bit Apple platforms
All modern Apple platforms (iOS, tvOS, watchOS, macOS) are 64-bit
and use ULong for NSStringEncoding. Updated the implementations
to use toULong() consistently.

+++ buffer/src/iosMain/kotlin/com/ditchoom/buffer/StringDecoding.ios.kt
@@ -10,7 +10,7 @@ import platform.Foundation.create
 
 /**
  * iOS implementation using Foundation's NSString.
- * On iOS, NSString.create expects UInt for encoding.
+ * On 64-bit iOS, NSString.create expects ULong for encoding.
  */
 internal actual fun decodeWithFoundation(
     data: ByteArray,
@@ -23,6 +23,6 @@ internal actual fun decodeWithFoundation(
             NSData.create(bytes = pinned.addressOf(offset), length = length.toULong())
         }
 
-    return NSString.create(nsData, encoding.toUInt())?.toString()
+    return NSString.create(nsData, encoding.toULong())?.toString()
         ?: throw IllegalArgumentException("Failed to decode bytes with encoding: $encoding")
 }
